Fixes after code review were implemented
