+++ src/components/CourseCard/index.ts
