+++ src/components/Icon/index.ts
