+++ src/components/Lesson/index.ts
