+++ src/components/RatingStars/RatingStars.tsx
@@ -1,4 +1,4 @@
-import { IconType } from '../../types/IconType';
+import { IconType } from '../../types/enums';
 import { Icon } from '../Icon';
 import './RatingStars.scss';
