+++ src/components/ToTopButton/ToTopButton.tsx
@@ -1,6 +1,6 @@
 import { Link, useLocation } from 'react-router-dom';
 
-import { IconType } from '../../types/IconType';
+import { IconType } from '../../types/enums';
 import { Icon } from '../Icon';
 import './ToTopButton.scss';
