--- src/types/ButtonType.ts
@@ -1,1 +1,0 @@
-export type ButtonType = 'prev' | 'next';
