--- src/types/Courses.ts
@@ -1,23 +1,0 @@
-export interface Courses {
-  id: string;
-  title: string;
-  tags: string[];
-  launchDate: string;
-  status: string;
-  description: string;
-  duration: number;
-  lessonsCount: number;
-  containsLockedLessons: boolean;
-  previewImageLink: string;
-  rating: number;
-  meta: {
-    slug?: string;
-    skills?: string[];
-    difficulty?: string;
-    courseVideoPreview?: {
-      link?: string;
-      duration?: number;
-      previewImageLink?: string;
-    }
-  };
-}
