--- src/types/ErrorType.ts
@@ -1,4 +1,0 @@
-export enum ErrorType {
-  GET_COURSES = 'Unable to get the courses',
-  GET_COURSE_DETAILS = 'Unable to get the course\'s details',
-}
