--- src/types/LessonType.ts
@@ -1,11 +1,0 @@
-export interface LessonType {
-  id: string;
-  title: string;
-  duration: number;
-  order: number;
-  type: string;
-  status: string;
-  link: string | URL;
-  previewImageLink: string;
-  meta: string | null;
-}
