--- src/types/LogoType.ts
@@ -1,1 +1,0 @@
-export type LogoType = 'big' | 'small';
